Updated openstack/openstack Project: openstack/oslo.config  0abac4881355a395a14a5d68e3347805a480cff9
Log a warning when deprecated opts are used

This change logs the warning at lookup time instead of parse time
because we have no way of knowing at parse time which options are
deprecated.  It does use a set to make sure it only logs a given
deprecation warning once per option so it won't spam the logs.

I could not find a way to log deprecations for cli args.  argparse
seems to only write the value to the undeprecated name, so there's
no way to tell whether it came from a deprecated name originally.
Since most of our opts are set through config files I don't see
this as a huge issue though.

I'm also not using the warnings module because its deprecation
mechanisms seem more developer focused, and for config opts we
need to notify the user.  This is not possible with the warnings
module in Python 2.6 (there's no way to send warnings to the log
by default), and since we still need to support that for oslo libs
I went with a simple logged warning instead.

Change-Id: I4580831f34c48882fb07288fa8d55272b58e2674
Closes-Bug: 1390408