Updated openstack/openstack Project: openstack-infra/puppet-ansible  1769fbd75dc96a77afa6b59c9bec3bff09361be5
Remove embedded ansible role

Once we have standalone ansible roles, we don't need to embed the
ansible code in the puppet module. We also don't need to delete the
existing roles path, as the follow on standalone role should go in
place, but we want the existing role to exist in the mean time.

Change-Id: I76d5cab55942beaff44ea5f289f93ff6ce772c5f