Updated openstack/openstack Project: openstack/python-designateclient  debf39a62981b3a5ad6eaf745c73f90af4971a67
Added extra previllege to list all domains from all tenants

The user has to use the command "designate --all-tenants domain-list" to list all domains from all tenants.
When the above command is used "X-Auth-All-Projects" value is set to true and it is passed as a header to
designate.This will allow us to list all domains from all tenants.

Change-Id: I4cd4dd5427f5f35cdec95dbdf36c7386b60a2949
Fixes: bug #1418156