Updated openstack/openstack Project: openstack/sahara-image-elements  de118b5ab1a7cf1ea25bb3bda37498b663b48739
Rename a couple of scripts to avoid duplicated names

In recent versions, diskimage-builder aborts when two elements have
scripts in the same hook with the same filename. Since hive and oozie
are usually used/required by the element of a plugin, there will be at
least 2 or 3 root.d scripts named 0-check.

Give unique names to the hive and oozie scripts.

Change-Id: Ia85508615de4271dfc1de913b7eff2d36392c008
Closes-Bug: #1424715