Updated openstack/openstack Project: openstack/oslo.versionedobjects  42baafb810827c35d0e6eaade49631e7bc88d14b
Allow subclasses to define the serialization namespaces

This allows the serialization and project namespaces to be overridden in
subclasses of VersionedObject. The project one should be set on each
using implementation, but the base serialization one is just there for
projects that already use this code and need it to be equal to what they
currently use.

Change-Id: Iecec043adc31441cc8865fa5e54d9a0ee4a5989a