Updated openstack/openstack Project: openstack/nova  cb287b45a93f1b69255952e930529dde354a31ba
Add methods that convert any volume BDM to driver format

There are places in the code where we need to get the DriverBlockDevice
instance based on the BDM object for a volume block device, however the
DriverBlockDevice class the object maps to (based on source of
the data) can be any of the following: volume, snapshot or image, all of
which are actually separate classes that extend VolumeDriverBlockDevice.

This patch adds a method to transform a volume BDM object to the correct
class instance, so that we can more easily avoid bugs such as 1411847.

Related-Bug: #1411847

Change-Id: I3496bab7aac448722240483b86c013e5f80d0cb5