Updated openstack/openstack Project: openstack/nova  2789cff1937e5eb8313129a2f339907bfda9d1d2
extract API fixture

Instead of doing inline setup of the API service in integrated_helpers
extract this functionality to a fixture so that it can easily be
reused in other functional tests.

The setup work to patch the global CONF to be able to run the API
server is also extracted into a fixture, and the inline setup in
integrated_helpers is removed.

fake_network=True is already set in the global conf_fixture used by
nova.test, setting it again here is redundant (also, aren't deep
inheritance stacks in tests fun‽)

The _get_flags support is still required because ~ 13 tests manipulate
the extension lists with it (though they make no other changes). We
should semantically extract that concept so that a test can declare an
extension list more abstractly, and we get that to the OSAPI fixture.

Change-Id: I4d8224291d729aee04049e3034499e5fe6c49f21