Updated openstack/openstack Project: openstack/nova  df787e7fe5a0e63fe8a79218934156a8fc0ff5a1
Change utils.vpn_ping() to return a Boolean

vpn_ping() is used exlusively to tell if a vpn server is actually
listening. The session_id it picks is not sensitive. It is just using
the session_id to make sure that the packet it receives back is a
response to the packet it sends so that it can measure latency and
tell that the server is up.

Also the tests in test_cloudpipe.py are already treating the return
value as Boolean.  Lets make it explicit.

Change-Id: I2a1d98b5fb3c9c02a4161c6df2473e091442d01d
Closes-Bug: #1420457