Updated openstack/openstack Project: openstack-dev/devstack  e2b95ff2a2ee2d91e31ad0c00babdeabcf1cecb3
F21 requires redhat-rpm-config for mariadb-devel

As you can see in the f21 job, the python-MYSQL compilation
fails without this library.

At the first look it seams like packaging issue, so
I created bug related to this. [1]
Until the issue is addressed, the redhat-rpm-config will be
considered as build time dev dependency for f21.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1195207

Change-Id: I731828cac912d1b11eaf2269364411c77a57d76c