Updated openstack/openstack Project: openstack/tempest  fbc45c70313184e42fe14484f6e1e044fd776549
Adding 'security_groups' attributes in server schema

If Nova 'os-security-groups' extension is enabled then, server post,
get & detail response will contain 'security_groups' attributes.

So that attributes should be defined in server's schema and it should
not be required as this is present only when sg extension is enabled.

This patch adds the 'security_groups' attributes in necessary schema.

Change-Id: Ibb58a2173171d237156d383ced9c3b58993058bb