Updated openstack/openstack Project: openstack/cinder  146e5d1ff7c9f0c36c1a806f628148348bc1cdba
Fixed 3PAR driver load balancing during migration

This patch will not allow the 3PAR iSCSI driver to pick another
iSCSI IP from the hp3par_iscsi_ips list during a nova live-migration.

Nova expects that the IQN will be exactly the same as it's attaching
the volume(s) to the new host during live-migration. This conflicts
with the Cinder settings such as "hp3par_iscsi_ips" which allows for
multiple IPs for the purpose of load balancing.

Change-Id: Idee2fdf7d5bbbe8fb7bdbc3ff20622a213738aa5
Closes-bug: 1423958