Updated openstack/openstack Project: openstack/taskflow  84e04ba6ea6230124cea55acbd0b3e9d33db8b30
Refactor parts of the periodic worker

Perform some adjustments to the periodic worker type to
make it easier to later change and make it better in the
future. These changes hopefully also reduce complexity
and increase understandability.

In part this deprecates providing a tombstone via
__init__() since that will restrict future enhancements we can
make (in retrospect we need more internal control over that
argument & type to build better workers).

Change-Id: I1965e157c303c2a45b9950e9f4a921c638f57fd1