Updated openstack/openstack Project: openstack/cinder  9fd5bb206bb2b72dde323ca75ced49080c0c5a82
Remove retry_deactivation directive on lvremove

A while back an LVM_CONF directive was added to
explicitly retry_deactivation on lvremove.  This
was a great idea, but since then the LVM defaults
all seem to set this for us anyway so we don't
really need to send it in the call explicitly
any more.

Also, we've introduced the option to use a cinder
specific lvm.conf file which can be used to set
or unset values like this going forward:
    (commit: eb7bb3e08c1b8e1008ba447b842b635821e2097e)

This patch removes the retry_deactivation directive, but
leaves the ignore_suspended_devices setting as we don't
want to use that all the time, but only during recovery.
There may be better options for this call as well but
that's a possible future improvement unrelated to this
commit.

Change-Id: I167ea9a24cfeeb3716c90c87a0609271d3255ca6