Updated openstack/openstack Project: openstack/neutron  d4dca6671afbecb640748c89e766ec0b7d2b0ebc
Default route missing for IPv6 subnets in HA Router

In an HA setup, keepalived would configure the default gateway (in the master
HA router) by parsing the "virtual_routes" section of keepalived.conf file.
The "virtual_routes" section is properly constructed for an IPv4 subnet, but
requires a minor change for IPv6 subnets.

Closes-Bug: #1414937
Change-Id: I54070e048577292d98d86fcc9ca8b9d3e72b4ebe