Updated openstack/openstack Project: openstack/nova  809eaca1790ec8386485fb9d020f33c07ab9a9c7
libvirt: address test comments for zfcp volume driver changes

This addresses the testing comments in change
Ica4938d9962bbb402e26dd47f0aea502ea0c4245.

Given the FFE deadline, the fact we have a +2 on the change, the code is
functionally OK and the timezones that the authors and reviewers are in,
this was better to do than rev the driver changes again.

Change-Id: I4eb949f12fd6eb2dd6a0d738a9d2e5468dc3c324