Updated openstack/openstack Project: openstack/horizon  d6523f834bdb1335a4d3dcba13b62fc4df548d7f
Go button on the Instance Log page should be enable all the time.

The "Go" button on the instance log page will be disable after user clicked on it. User needs to refresh the page again to make the button enable.
 
Steps to reproduce.

1. Spin up an instance.
2. Once the instance is up, click the View Log action.
3. On the view log page, enter some valid value on the Log Length text box.
4. Click "Go" button.
5. Once the page is refreshed, you will notice that the Go button become disable forever.

The root cause is due to Horizon button will be disabled once user submitted a form.
This behavior works great with regular form submission such as create an instance etc.

However view log page seems to be a single page application. That is, browser does not need to refresh the page to view the log with updated log length. Therefore, we need to mark the "Go" button
to be enabled all the time.

Change-Id: I1e8d582595eb6825109709bd2e31bfb6f1e6d6ba
Closes-Bug: #1424244