Updated openstack/openstack Project: openstack/cinder  5924a31be3beaae58c19baf02814da9a68a511ed
FCZM fix reading of cinder config entries

The zone manager and the lookup service had some cruft code that
was incorrectly reading the cinder.conf sections it needed.
When cinder.conf was configured to use the Cisco driver, the zone
manager would never see it, because it was reading from the wrong
section of cinder.conf.

Change-Id: I4a48bc39e55d4b7e9a149a3ec3aa3947980360ea
Closes-Bug: 1424852