Updated openstack/openstack Project: openstack-dev/devstack  5c053a6c569e3dd9608e95c012b2d37281536fe2
Add debtcollector to lib/oslo and stackrc

Ensure that the debtcollector library is pulled in
like the other oslo libraries so that devstack can
use it where appropriate.

Also fixes 'test_libs_from_pypi.sh' to not have a huge
single line of libraries; and splits it into multiple
lines so the diffs and code can be easily looked at.

Change-Id: I35ab0ed0e20b6092a41ecb3b6f1aaf0a05f5180e