Updated openstack/openstack Project: openstack/nova  60c817838900b1181eeb7947aee0c72993cc214f
libvirt: Fix migration when image doesn't exist

When nova is used without shared_storage and glance doesn't have the
images required by the instance anymore, the live block migration code
can't prepare the destination host properly.

This patch catches the case when images are not found in glance, and
copies the missing images from the source host, like the
migrate_disk_and_power_off code already does.

The KVM disk deep copy method is not used because it won't work for AMI
image that depends of AKI/ARI image. kernel and initrd are not considered
by the kvm disk migration because they are readonly file (by the kvm pov)
used only to boot the VM.

Co-Authored-By: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@cloudwatt.com>
Closes-bug: #1270825

Change-Id: If81f8b1bbe3e738579ffe2d8f36807afb77560d8