Updated openstack/openstack Project: openstack/nova  d27f7c41bf6c4a2586473db97f4931445cf41629
Make pagination work with deleted marker

Currently, Nova does not paginate correctly if the marker is a
deleted server. When trying to get servers with deleted and marker,
a "BadRequest" error saying that the marker is not found will be
raised.

Fixes the bug by passing read_deleted to context when getting
server to enable to get deleted server as marker.

Change-Id: Icde4e7d251d4497ee8eeb0a573eca3b73aa71466
Closes-bug: #1398086