Updated openstack/openstack Project: openstack/nova  0470a1dfc6aa50d21d91114cf342391c5a3cac35
Ignore 'dynamic' addr flag on gateway initialization

When initializing the gateway device, if 'dynamic' flag is present
but not ignored, the 'ip addr del' command will fail and nova-network
will fail to startup.
In LinuxBridgeInterfaceDriver, it is correctly ignored.
This adds the same handling of flags in initialize_gateway_device.

Change-Id: Id03e9659a5bec549d6b119cd0ced595875baf0c9
Closes-Bug: #1408142
Related-Bug: #1274611