Updated openstack/openstack Project: openstack/nova  8f64485a1f3b75b75100f90f9307952e30abbdfa
Fix BM nodes extension to deal with missing node properties

The os-baremetal-nodes extension currently relies on the presence
of specific properties on the Ironic nodes.  This can result in
KeyError exceptions if there are nodes in the Ironic inventory
without these (specifically: cpus, memory_mb, local_gb).  This
updates the extension to avoid the KeyError. This is consistent
with how the Ironic driver calculates resources from nodes.  That
is, nodes that are missing a specific property are simply considered
to provide 0 resources of that type.

Change-Id: I1c30d5100b01a3e8cbce5185b5adeb8b5ce48aa0
Closes-Bug: #1423427