Updated openstack/openstack Project: openstack/tempest  efb1985355525fbfc156dcfaa204ecee908d497a
Fix AttributeError on testtools missing MismatchError

It sounds like MismatchError is borderline between part
of a public interface of testtools, and may or may not be
available depending of testtoools version.

Get rid of the nonsense altogether. The helper method was
wrong in that it shouldn't have been asserting a result if an
expected condition was not that. This commit removes the
broken helper method in favor of directly calling the 
get_dns_servers method where necessary.

Closes-bug: #1424482

Change-Id: Iaa611f833f70a528f38fe8cc9d9add0047bf51e1