Updated openstack/openstack Project: openstack/oslo.db  7bfdb6a704855984ae35a1c6ef063782a4f7bf1d
Make DBAPI class work with mocks correctly

Currently, DBAPI __getattr__() implementation will *always*
implicitly wrap all mocked methods of a target class/module using
wrap_db_retry(), as retry_on_* attributes will always exist and be
true for mock instances. Not only this leads to a situation, when
we would retry methods, we were not going to retry, but this also
fails with AttributeError on functools.wraps() call, breaking valid
unit tests in projects.

DBAPI is changed in a way, so that it does dict look ups instead of
fetching instance attributes to ensure mocked methods are handled
correctly and aren't wrapped when they shouldn't be wrapped.

Related issue:
http://stackoverflow.com/questions/22204660/python-mock-wrapsf-problems

Change-Id: I36550ffa88056c53a431d64e61e84fbb44bbf68d