Updated openstack/openstack Project: openstack/python-swiftclient  925c01ebfbdfb6478a3786f24a9572deae40f8f8
Release connection after consuming the content

When using stream=True, we need to make sure we release a connection
back to the connection pool to avoid warnings clogging up the logs. We
can do this by releasing the connection once we've read all the content
from the response. In situations where stream=False, requests already
does this for us.

Related-bug: 1424732
Closes-bug: 1341777
Change-Id: Id1c72ad86135a49d4b985992d788736b65f7dbda