Updated openstack/openstack Project: openstack/tempest-lib  4c4d34b2b0d70e0c92f568a0080eff0df49f096f
Handle missing launchpadlib gracefully

The use of launchpadlib should not be required for tempest-lib, however
because we wanted to expose the skip_tracker as an entry point and
package it with the library an import to launchpadlib was needed in the
packaged source, which normally should require an entry in the
requirements file. This commit addresses that by making the
launchpadlib use not mandatory. The script will check if launchpadlib
is available and if it isn't the skip tracker will just exit after it
scans for the bug list.

Change-Id: I59814346e899e366734caaf757eb2d8fe5ecbcf7