Updated openstack/openstack Project: openstack/tempest  d93e79cd979aa303ff8ee69c1aed63cf88422973
Fix missing variable in test accounts warning log

In the remove_hash test accounts method we log a warning message to
indicate that a hash has already been cleaned and the pseudo-lock file
isn't there. The intent of that log message was to indicate the lock
file path in the warning. However the string never passed the hash_path
variable so instead %s was being printed. This commit fixes the issue
so the proper path will be in the warning log.

Change-Id: I87b66e3222452200996020c349cf05cb842114bf