Updated openstack/openstack Project: openstack/nova  83bd6b4fbdff2037278f4ac095e7741063d195b6
Better logging of resources

Consolidate a bunch of logging messages into one when syncing resources.
This reduces a lot of log messages for virt drivers that manage a lot
of nodes within the same nova-compute. This also now logs the node name
with the resources, so you can actually tell which node we're talking
about.

Change-Id: I0ee9c9c5913d48b96f70250ed5540424d88c083b