Updated openstack/openstack Project: openstack/ironic  b553f485fe5559187db3d74918b4e200878432e1
Root partition should be bootable for localboot

When instance are booted with 'boot_option' as 'local', Ironic
installs boot loader (that is present within the image) and then
makes the machine to boot from disk. This will fail on some BIOS
if the root partition is not marked as bootable.  This commit marks
root partition as bootable when boot_option is 'local'.

Change-Id: Ic88b4dda2d73be54fde5046f85356aba641a0fbe
Closes-Bug: 1422721