Updated openstack/openstack Project: openstack/tempest  823981a2b45a30448138896189d7e5226cebff3a
Revert "Change messaging clients to return one value and update tests"

This change broke Zaqar's gate. Please, lets revert it and re-try adding
it with a proper fix. I believe the issue is that it expects to receive
a dictionary everytime, whereas some of Zaqar's responses are lists.

Please see this review:
https://review.openstack.org/#/c/151529/6 (mongodb job)

This reverts commit 4b51340639a5b5f876fdc29c494bccf4a6dc5b6a.

Change-Id: Ie89c988b2a06b557238f9a15616bb4a8bb5b2a17