Updated openstack/openstack Project: openstack/cinder  03139fc441dcc3ace395804f0c54fedd7a1c1be5
Use Unique SCST Group names in SCST cinder helper driver.

Currently, while setting up GROUPS under targets in scst.conf,
we are using the Initiator IQN as the group name. When we have
multiple targets and we want to connect LUN's from these multiple
targets to the same initiator, we are using the same GROUP name in
all the Target sections.

It would be good to provide a unique group name by prepending the
GROUP name with the target name. So the GROUP name will be
"Target name+Initiator IQN".

Change-Id: Ia0fbab18f792e602e1f4dae6f9ca0e607d0cfde3
Closes-Bug: #1424973