Updated openstack/openstack Project: openstack/python-keystoneclient  3e2035bf8823f87cb736a526d4f084487066d33e
Allow handling multiple service_types

If the same service_type was mentioned in the catalog more than once
then only the last entry would be parsed and any possible other matches
would be lost.

This was something that novaclient used to do, and as we are pushing
sessions as the way that clients should all work we need to maintain
that compatibility.

Change-Id: I6964515ed1975bce1998897abfc02a1ec36e2584
Closes-Bug: #1425766