Updated openstack/openstack Project: openstack/nova  51d748ad9e5bc757a558a6449556f1e2d440e952
Use flavor stored with instance in libvirt driver

Currently, the libvirt driver does a Flavor.get_by_id() to retrieve
the flavor in order to generate the guest config, if a flavor was not
passed in. This results in a query for the flavor from the
instance_types table, which can fail in a cells environment
(FlavorNotFound errors in the check-tempest-dsvm-cells job), as flavors
are not synced from the top cell to child cells in any way.

Since the flavor-from-sysmeta-to-blob patches merged, the instance
flavor is available from the instance itself. This change replaces
calls of Flavor.get_by_id() in the libvirt driver with accesses of
the instance.flavor attribute.

Related to blueprint kilo-objects

Change-Id: I7fba490710e32a4cfd7615df672136d9849358ec