Updated openstack/openstack Project: openstack/keystone  143c3fedca5a49deef5fab641c3ada6ff628e6ec
Remove check_role_for_trust from sample policies

The "identity:check_role_for_trust" was defined in the sample
policy files but there is no actual mapping for it, so setting
a value for this target has no effect. If or when the
mapping gets added then this target must be added back in.

Fixed the double protected call in "get_role_for_trust" by changing its
call to a private unprotected version of "check_role_for_trust".

Also, marking the public version of "check_role_for_trust" as deprecated
for future cleanup.

Change-Id: I1c2b1186e37e31eaf556f81db686cc362768a5ae
Closes-Bug: #1421966