Updated openstack/openstack Project: openstack-infra/shade  eea3e20120bbe4ecad71dd7f0dd4425f4c0ce7be
Utilize dogpile.cache for caching

In the future we can make the backend more configurable which will help
users build more powerful things on top of shade with better caching
available to them.

Shade is caching elements like images, volumes and flavors
that are subject to change (flavor the less of them). Storing
items in a cache makes that changes to elements, like adding
or deleting, or changes of status, are not perceived by the
cache. However, the cache is still needed to improve performance
when several concurrent calls are done.

Change-Id: If072bcdbdaa2b0a021e26b1ea2211b1dfcc75880
Co-Authored-By: Clint Byrum <clint@fewbar.com>