Updated openstack/openstack Project: openstack-infra/elastic-recheck  d4fc4206539774fc43b5b89a0adb5912516208f3
Add query for cinder bug 1425310

This adds the new issue we're seeing with test_rescue
and also cleans up the bad query we have for bug 1373513.

It'd be awesome if we could only hit this when a test fails
but it seems we can't query on collections of items from
multiple log files.  This is kind of annoying because we
can have the slow LVM condition but not actually fail for
anything.  In fact the rescue detach test will almost
100% of the time result in this message being in syslog,
but we still make it within tempest's 196 seconds and the
test passes.  But, if some other "thing" fails it gets
attributed to this query.

Change-Id: I55da9d25e2ee8a79aba66d49dd922e7709188063