Updated openstack/openstack Project: openstack/cinder  0646a7d388afafb00894eff97e8ae58187ff70ab
Replication status periodic task optimization

A periodic task exists to update the replication status for all volumes.
Currently, this task executes for all drivers and always retrieves all
volumes for the current host from the DB.

This patch set:
* Ensures that the periodic task is only activated if the driver actually
  supports replication
* Only retrieves volumes from the DB if 'replication_status'!='disabled' in
  the periodic task

Also, the driver documentation in cinder.volume.driver.VolumeDriver.
get_volume_stats() is updated to reflect that the 'replication' key indicates
that the driver supports replication; this is the key that was actually
implemented in the drivers that support replication.

Change-Id: I61fbc31567ad0b6908a00113adeaccf415343e8e
Closes-Bug: 1383524