Updated openstack/openstack Project: openstack-infra/project-config  9a0c892d5cf4ff9311fdc9f6f38ed2022dc42b1e
Use neutron's api job to validate func job change

A recent commit to devstack-gate added support for limiting the repos
that are updated for a node by overriding the PROJECTS variable.  This
change temporarily repurposes neutron's api job to check that setting
this override is compatible with the functional job.  The api job is
currently non-voting, so no patches will be harmed in the process.

Change-Id: I7591569d3f039eeebed859333b29dd63104cd389