Updated openstack/openstack Project: openstack-infra/zuul  58fbc8854a5beb30b6b46920e3bc655749bbbe75
Don't report on non-live changes dequeueing

If a non-live change was removed from the queue because one of
its dependencies was removed, do not report on that change.  The
live change will eventually be removed for the same reason and
it will report.  Once is all that is needed.

This should not normally happen, but we have seen it occur when
dependencies have somehow changed behind zuul's back, possible
because of an error in Gerrit.

Change-Id: Ibb8daafb064e841606083d648b81389966d5333e