Updated openstack/openstack Project: openstack/ironic  4c318e107b05797e0a6dde69e033cdc4e5c8fe7e
Update documentation on VirtualBox drivers

This has minor cleanup of the documentation on VirtualBox
drivers. It fixes some grammatical issues and adds a bit more
clarification to some of the instructions for setting up a
developer's environment.

This is a follow up to I2018f9a89a3644863569ddac3126db71b125fb59.

Change-Id: I0f84554c98317160bfdfee14ba923cff4adda773
Implements: blueprint ironic-virtualbox-webservice-support